Added chirping to usertimeline
