+++ src/Chirp.Web/Pages/UserTimeline.cshtml
@@ -7,6 +7,17 @@
 }
 
 <div>
+    @if (User.Identity.IsAuthenticated)
+    {
+    <div class="cheepbox">
+        <h3>What's on your mind @(Model.Username)?</h3>
+        <form method="post">
+            <input required style="float: left" type="text" asp-for="cheepDTO.Text" maxlength="160">
+            <input type="submit" value="Share">
+        </form>
+    </div>
+    }
+    
     <h2> @routeName's Timeline </h2>
 
     @if (Model.Cheeps.Any())
